Fix promises clearing parameters
This moves the control flow handling after returning from a promise into
the sequencer. This is a change I really wish I could've made in 2017-it
reaches into the heart of the sequencer and changes some ancient code,
but all the tests pass so it doesn't *seem* to have subtly broken
anything.

It looks like reuseStackForNextBlock is no longer a mere optimization--
unlike popping and re-pushing, it preserves stack frame values, which we
rely on because we set the parameters on the top stack frame, whereas
before, we would initialize them and then push another stack frame. I
could also change procedures_call back to initializing the params then
pushing another stack frame, but this seems to work fine.